requirements.txt

fontawesomefree==6.5.1
django-bootstrap-v5